compilation error for HP-UX fixed
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/branches/WX_2_2_BRANCH@6973 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/gtk/dcclient.cpp
@@ -1284,8 +1284,8 @@ void wxWindowDC::DoDrawRotatedText( const wxString &text, wxCoord x, wxCoord y, 
         for ( wxCoord srcY = 0; srcY < h; srcY++ )
         {
             // transform source coords to dest coords
-            double r = sqrt(srcX*srcX + srcY*srcY);
-            double angleOrig = atan2(srcY, srcX) - rad;
+            double r = sqrt((double)srcX*srcX + srcY*srcY);
+            double angleOrig = atan2((double)srcY, (double)srcX) - rad;
             wxCoord dstX = (wxCoord)(r*cos(angleOrig) + 0.5),
                     dstY = (wxCoord)(r*sin(angleOrig) + 0.5);
